Initial Work on AI Manager.
Currently, the AI Manager is fully functional using Ollama to interact with an LLM. To create the manager, we can use a model file or a default llama3 model will be used instead. We are able to save entire conversations as well and the model can pick up where it left off if a conversation history JSON is provided.